requirements.txt
google-cloud-vision==3.4.4
opencv-python==4.8.0.74
requests==2.31.0
